:recycle: [refactor] : querydsl 문 수정.
스칼라 서브쿼리말고 left join으로 묶어서 사용했습니다.
이유
- 그거 뭐죠 그거 전략엔티티 안에 총 구독수가 있더라고요. 그래서 그거 사용해서 총 구독수 줬습니다.

+++ src/main/java/com/investmetic/domain/user/dto/response/TraderProfileDto.java
@@ -16,12 +16,12 @@ public class TraderProfileDto {
     private String nickname;
     private String imageUrl;
     private long strategyCount;
-    private long totalSubCount;
+    private int totalSubCount;
 
 
     @QueryProjection
     public TraderProfileDto(long userId, String userName, String nickname, String imageUrl, long strategyCount,
-                            long totalSubCount) {
+                            int totalSubCount) {
         this.userId = userId;
         this.userName = userName;
         this.nickname = nickname;
